requirements.txt
mysql-connector-python
pyspark==3.3.4
kafka-python
python-dotenv
pymongo
redis
findspark
avro-python3
fastparquet
pymysql
cryptography